requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
alembic
python-dotenv
jinja2
pydantic
